Adds room create and list functionality
Adds room description to room data and API.

Implements displaying a list of rooms with data from the database, including name, description, and member count.

Adds UI elements for creating a room with a name and description.

Fixes an issue where the test database file would not be ignored by git depending on the build target.