Adds contacts manager feature
Adds a contact manager feature to desktop and room dash for mobile app, including UI elements and routing.

The desktop app now includes a contacts manager accessible from the desktop navigation.
The mobile app dashboard has been updated and refactored to adhere to the new design system and prepare for the addition of a contacts feature in the future.

Includes CSS styling, internationalization and data fetching.