Fix test failure, formatting, and clippy signal error
The icons no longer have the same sizing functionality because of use if inline svgs. I need to decide how best to handle that. Really the right way would be to use the size values I want in the inline svg values. The way its loading the svg strings currently is a little strange anyway and I'll likely go with a cleaner solution at some point.